+++ internal/http/utils/errorpage.go
@@ -13,12 +13,35 @@ type templatePicker interface {
 	Pick() string
 }
 
-func RespondWithErrorPage(ctx *fasthttp.RequestCtx, cfg *config.Config, p templatePicker, code string, httpCode int) {
+func RespondWithErrorPage( //nolint:funlen
+	ctx *fasthttp.RequestCtx,
+	cfg *config.Config,
+	p templatePicker,
+	code string,
+	httpCode int,
+	showRequestDetails bool,
+) {
+	ctx.Response.Header.Set("X-Robots-Tag", "noindex") // block Search indexing
+
+	if returnCode, ok := extractCodeToReturn(ctx); ok {
+		code, httpCode = strconv.Itoa(returnCode), returnCode
+	}
+
 	var (
-		_, canJSON = cfg.Formats[config.FormatJSON]
-		props      = tpl.Properties{Code: code}
-		clientWant = ClientWantFormat(ctx)
+		clientWant    = ClientWantFormat(ctx)
+		json, canJSON = cfg.JSONFormat()
+		xml, canXML   = cfg.XMLFormat()
+		props         = tpl.Properties{Code: code, ShowRequestDetails: showRequestDetails}
 	)
+
+	if showRequestDetails {
+		props.OriginalURI = string(ctx.Request.Header.Peek(OriginalURI))
+		props.Namespace = string(ctx.Request.Header.Peek(Namespace))
+		props.IngressName = string(ctx.Request.Header.Peek(IngressName))
+		props.ServiceName = string(ctx.Request.Header.Peek(ServiceName))
+		props.ServicePort = string(ctx.Request.Header.Peek(ServicePort))
+		props.RequestID = string(ctx.Request.Header.Peek(RequestID))
+	}
 
 	if page, exists := cfg.Pages[code]; exists {
 		props.Message = page.Message()
@@ -41,14 +64,27 @@ func RespondWithErrorPage(ctx *fasthttp.RequestCtx, cfg *config.Config, p templa
 	switch {
 	case clientWant == JSONContentType && canJSON: // JSON
 		{
-			SetClientFormat(ctx, clientWant)
+			SetClientFormat(ctx, JSONContentType)
 
-			if content, err := tpl.RenderJSON(cfg.Formats[config.FormatJSON].Content(), props); err == nil {
+			if content, err := tpl.Render(json.Content(), props); err == nil {
 				ctx.SetStatusCode(httpCode)
 				_, _ = ctx.Write(content)
 			} else {
 				ctx.SetStatusCode(fasthttp.StatusInternalServerError)
 				_, _ = ctx.WriteString("cannot render JSON template: " + err.Error())
+			}
+		}
+
+	case clientWant == XMLContentType && canXML: // XML
+		{
+			SetClientFormat(ctx, XMLContentType)
+
+			if content, err := tpl.Render(xml.Content(), props); err == nil {
+				ctx.SetStatusCode(httpCode)
+				_, _ = ctx.Write(content)
+			} else {
+				ctx.SetStatusCode(fasthttp.StatusInternalServerError)
+				_, _ = ctx.WriteString("cannot render XML template: " + err.Error())
 			}
 		}
 
@@ -59,7 +95,7 @@ func RespondWithErrorPage(ctx *fasthttp.RequestCtx, cfg *config.Config, p templa
 			var templateName = p.Pick()
 
 			if template, exists := cfg.Template(templateName); exists {
-				if content, err := tpl.RenderHTML(template.Content(), props); err == nil {
+				if content, err := tpl.Render(template.Content(), props); err == nil {
 					ctx.SetStatusCode(httpCode)
 					_, _ = ctx.Write(content)
 				} else {
@@ -72,4 +108,18 @@ func RespondWithErrorPage(ctx *fasthttp.RequestCtx, cfg *config.Config, p templa
 			}
 		}
 	}
+}
+
+func extractCodeToReturn(ctx *fasthttp.RequestCtx) (int, bool) { // for the Ingress support
+	var ch = ctx.Request.Header.Peek(CodeHeader)
+
+	if len(ch) > 0 && len(ch) <= 3 {
+		if code, err := strconv.Atoi(string(ch)); err == nil {
+			if code > 0 && code <= 599 {
+				return code, true
+			}
+		}
+	}
+
+	return 0, false
 }