+++ internal/http/utils/headers.go
@@ -1,0 +1,27 @@
+package utils
+
+const (
+	// FormatHeader name of the header used to extract the format
+	FormatHeader = "X-Format"
+
+	// CodeHeader name of the header used as source of the HTTP status code to return
+	CodeHeader = "X-Code"
+
+	// OriginalURI name of the header with the original URL from NGINX
+	OriginalURI = "X-Original-URI"
+
+	// Namespace name of the header that contains information about the Ingress namespace
+	Namespace = "X-Namespace"
+
+	// IngressName name of the header that contains the matched Ingress
+	IngressName = "X-Ingress-Name"
+
+	// ServiceName name of the header that contains the matched Service in the Ingress
+	ServiceName = "X-Service-Name"
+
+	// ServicePort name of the header that contains the matched Service port in the Ingress
+	ServicePort = "X-Service-Port"
+
+	// RequestID is a unique ID that identifies the request - same as for backend service
+	RequestID = "X-Request-ID"
+)